+++ hunmin/__init__.py
@@ -24,7 +24,7 @@ Supported languages (14):
     ja, zh, ko
 """
 
-__version__ = '3.40.0'
+__version__ = '3.40.1'
 __author__ = 'Hunmin Project'
 
 from .api import (